requirements.txt
streamlit
pandas
scikit-learn
plotly
seaborn
openai
python-dotenv
streamlit
openai
fpdf